+++ Tests/Helpers/AccessToken+Test.swift
@@ -1,0 +1,5 @@
+import XpringKit
+
+extension AccessToken {
+    public static let testAccessToken = "password"
+}