+++ Tests/Helpers/BearerToken+Test.swift
@@ -1,5 +1,0 @@
-import XpringKit
-
-extension BearerToken {
-    public static let testBearerToken = "password"
-}