+++ Tests/Helpers/Io_Xpring_LedgerSequence+Test.swift
@@ -1,7 +1,0 @@
-import XpringKit
-
-extension Io_Xpring_LedgerSequence {
-  static let testLedgerSequence = Io_Xpring_LedgerSequence.with {
-    $0.index = 12
-  }
-}