+++ XpringKit/ILP/AccessToken.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+/// A token used to authenticate to a connector account
+public typealias AccessToken = String