+++ XpringKit/ILP/BearerToken.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-/// A token used to authenticate to a connector account
-public typealias BearerToken = String