+++ XpringKit/ILP/IlpClient.swift
@@ -15,13 +15,13 @@ public class IlpClient {
     ///
     /// - Parameters:
     ///     -  accountID The account ID to get the balance for.
-    ///     -  bearerToken Authentication bearer token.
+    ///     -  accessToken Authentication access token. Can not start with "Bearer "
     /// - Returns: An AccountBalance with balance information of the specified account
     /// - Throws: An error If the given inputs were invalid, the account doesn't exist, or authentication failed.
     public func getBalance(for accountID: AccountID,
-                           withAuthorization bearerToken: BearerToken
+                           withAuthorization accessToken: AccessToken
     ) throws -> AccountBalance {
-        return try decoratedClient.getBalance(for: accountID, withAuthorization: bearerToken)
+        return try decoratedClient.getBalance(for: accountID, withAuthorization: accessToken)
     }
 
     /// Send a payment from the given accountID to the destinationPaymentPointer payment pointer
@@ -30,15 +30,15 @@ public class IlpClient {
     ///         Payment status can be checked in PaymentResult.successfulPayment
     /// - Parameters:
     ///     -  paymentRequest: A PaymentRequest with options for a payment
-    ///     -  bearerToken : auth token of the sender
+    ///     -  accessToken : access token of the sender. Can not start with "Bearer "
     /// - Returns: A PaymentResult with details about the payment.
     /// - Throws: An error If the given inputs were invalid.
     public func sendPayment(_ paymentRequest: PaymentRequest,
-                            withAuthorization bearerToken: BearerToken
+                            withAuthorization accessToken: AccessToken
     ) throws -> PaymentResult {
         return try decoratedClient.sendPayment(
             paymentRequest,
-            withAuthorization: bearerToken
+            withAuthorization: accessToken
         )
     }
 }