+++ XpringKit/ILP/XpringIlpError.swift
@@ -1,6 +1,0 @@
-/// Errors that may occur when interacting with Xpring's ILP infrastructure
-public enum XpringIlpError: Error {
-
-    /// The requested functionality is not yet implemented.
-    case unimplemented
-}