+++ XpringKit/ILP/auth/IlpCredentials.swift
@@ -9,21 +9,22 @@ internal class IlpCredentials {
     /// Instead, it is wrapped in IlpCredentials
     private var metadata: Metadata
 
+    private let BEARER_PREFIX = "Bearer "
+
     /// Initialize a new IlpCredentials
     /// self.metadata will be initialized and an Authorization header will be added to it.  The value
     /// of that entry will be the bearerToken with "Bearer " prefix.
-    public init(_ bearerToken: String) throws {
-        self.metadata = Metadata()
-        try metadata.add(key: "authorization", value: self.applyBearer(bearerToken))
-    }
-
-    /// Prepends 'Bearer ' to an auth token, if it is not already prefixed with "Bearer "
     ///
     /// - Parameters:
-    ///     - token: An auth token that either does or does not have a "Bearer " prefix
-    /// - Returns: A bearer token as a String, with a "Bearer " prefix
-    private func applyBearer(_ token: String) -> String {
-        return token.starts(with: "Bearer ") ? token : "Bearer " + token
+    ///     - accessToken: An access token with no "Bearer " prefix
+    /// - Throws: if accessToken starts with "Bearer "
+    public init(_ accessToken: String) throws {
+        if accessToken.starts(with: BEARER_PREFIX) {
+            throw IlpError.invalidAccessToken
+        }
+
+        self.metadata = Metadata()
+        try metadata.add(key: "authorization", value: BEARER_PREFIX + accessToken)
     }
 
     /// Get the Metadata to pass to network calls