+++ XpringKit/JavaScript/JavaScriptUtils.swift
@@ -30,7 +30,7 @@ internal class JavaScriptUtils {
 
   /// Initialize a JavaScriptUtils object.
   public init() {
-    utilsClass = XRPJavaScriptLoader.load(ResourceNames.Classes.utils)
+    utilsClass = JavaScriptLoader.load(ResourceNames.Classes.utils)
   }
 
   /// Check if the given address is a valid XRP address.
@@ -98,9 +98,9 @@ internal class JavaScriptUtils {
       return nil
     }
 
-    let classicAddress = XRPJavaScriptLoader.load(ResourceNames.Properties.address, from: result).toString()!
-    let isTest = XRPJavaScriptLoader.load(ResourceNames.Properties.test, from: result).toBool()
-    if let tag = XRPJavaScriptLoader.failableLoad(ResourceNames.Properties.tag, from: result) {
+    let classicAddress = JavaScriptLoader.load(ResourceNames.Properties.address, from: result).toString()!
+    let isTest = JavaScriptLoader.load(ResourceNames.Properties.test, from: result).toBool()
+    if let tag = JavaScriptLoader.failableLoad(ResourceNames.Properties.tag, from: result) {
       return (classicAddress, tag.toUInt32(), isTest)
     }
     return (classicAddress, nil, isTest)