+++ XpringKit/XRPPayment.swift
@@ -1,5 +1,7 @@
 import Foundation
 
+/// A payment on the XRP Ledger
+/// - SeeAlso: https://xrpl.org/payment.html
 // TODO(keefertaylor): Modify this object to use X-Address format.
 public struct XRPPayment: Equatable {
   public let amount: XRPCurrencyAmount